+++ src/graph/node_id.rs
@@ -1,4 +1,4 @@
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Copy, Default, Eq, Hash, PartialEq, Debug, Serialize, Deserialize)]
-pub struct NodeID(pub usize);
+pub struct NodeID(pub u64);
