+++ src/track/audio_track/mod.rs
@@ -29,7 +29,7 @@ pub struct AudioTrack {
     audio_ctx: AudioContext,
 
     // --- MISC ---
-    next_region_id: usize,
+    next_region_id: u64,
 }
 
 impl AudioTrack {
@@ -66,7 +66,7 @@ impl AudioTrack {
 
     // --- REGION ADDITION ---
 
-    pub fn set_next_region_id(&mut self, next_id: usize) {
+    pub fn set_next_region_id(&mut self, next_id: u64) {
         self.next_region_id = next_id;
     }
